+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Dynamic find with empty types or mimes
 
+### Fixed
+- Banner classification list (correct medium)
+
 ## [2.3.2] - 2023-01-31
 ### Fixed
 - Calculating waiting payments